% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrMat_function.R
\name{CorrMat}
\alias{CorrMat}
\title{Specifying a correlation matrix for species.}
\usage{
CorrMat(n, groups = NULL, values = NULL)
}
\arguments{
\item{n}{Dimension of the correlation matrix}

\item{groups}{List of indices that belong to the same correlation class. The
correlation among indices in the same element of the list will be non-zero.}

\item{values}{Vector of correlation values for the indices belonging to the
different elements in the list groups.}
}
\value{
Correlation matrix.
}
\description{
Specifying a correlation matrix for species.
}
