% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_function.R
\name{MCMC}
\alias{MCMC}
\title{MCMC for bipartite network model with trait information and unrecorded
interactions}
\usage{
MCMC(
  obs_A,
  obs_n,
  obs_X,
  obs_W,
  Cu,
  Cv,
  Nsims,
  burn,
  thin,
  use_H = 10,
  bias_cor = TRUE,
  theta_inf = 0.01,
  mh_n_pis = 100,
  mh_n_pjs = 100,
  mh_n_rho = 100,
  stick_alpha = 5,
  prior_theta = c(1, 1),
  prior_tau = c(5, 5),
  prior_rho = c(5, 5),
  prior_mu0 = 0,
  prior_sigmasq0 = 10,
  prior_sigmasq = c(1, 1),
  start_values = NULL,
  sampling = NULL
)
}
\arguments{
\item{obs_A}{Observed interaction matrix. Contains values of 0 and 1.}

\item{obs_n}{Matrix with the same dimensions as obs_A. Includes the number
of studies in which both species were observed.}

\item{obs_X}{Matrix of observed covariates for the first set of species.
Rows correspond to species, columns to covariates. Continuous covariates
should be first, and binary covariates should follow.}

\item{obs_W}{Same structure as obs_X but for the second set of species.}

\item{Cu}{Phylogenetic correlation matrix for the first set of species.}

\item{Cv}{Phylogenetic correlation matrix for the second set of species.}

\item{Nsims}{Number of posterior samples to be kept.}

\item{burn}{Number of samples to be burnt in the beginning of the MCMC.}

\item{thin}{MCMC sampling thinning.}

\item{use_H}{Number of latent factors to be used. Defaults to 10.}

\item{bias_cor}{Logical. Whether the model should aim to correct for
geographical and taxonomical bias. If set to FALSE, the MCMC is simplied
since the observed matrix of interactions is considered the same as the
true matrix of interactions. Defaults to TRUE.}

\item{theta_inf}{Variance value theta infinity in the spike part of the
prior distribution for increasing shrinkage. Defaults to 0.01.}

\item{mh_n_pis}{Parameter n in the Beta proposal for updating pi. Defaults
to 100.}

\item{mh_n_pjs}{Same as mh_n_pis but for the update of pj.}

\item{mh_n_rho}{Same as mh_n_pis but for the update of rho.}

\item{stick_alpha}{Alpha value in the increasing shrinkage prior. Defaults
to 5.}

\item{prior_theta}{Hyperparameters of the inverse gamma distribution in the
slab part of the increasing shrinkage prior. Defaults to (1, 1).}

\item{prior_rho}{Hyperparameters of the beta prior for the weight rho in 
the correlation matrix for latent factors. Defaults to (5, 5).}

\item{prior_mu0}{Mean of the normal prior distribution for all intercepts.
Defaults to 0.}

\item{prior_sigmasq0}{Variance of the normal prior distribution for all
intercepts. Defaults to 10.}

\item{prior_sigmasq}{Hyperparameters of the inverse gamma prior on the
variance terms of continuous traits. Defaults to (1, 1).}

\item{start_values}{List that can include starting values for any subset of
parameters. Starting values for parameters that are not specified in
start_values will be sampled. Defaults to NULL.}

\item{sampling}{List specifying which parameters should be sampled by
setting the value to TRUE, and which should not and should be kept at
their starting value by setting the corresponding list element to FALSE.
If sampling is set to FALSE for a parameter, it is recommended that the
corresponding start_value is specified to be set to the parameter's true
value. Defaults to NULL, and when set to NULL all parameters are sampled.}

\item{pruir_tau}{Hyperarameters of the inverse gamma prior for the
additional variance parameter in the coefficients. Defaults to (5, 5).}
}
\description{
Perform MCMC to acquire samples from the posterior distribution of model
parameters for a model for the true interaction matrix among different sets
of species with available trait information.
}
