% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraitMatching_function.R
\name{TraitMatching}
\alias{TraitMatching}
\title{Trait matching for species interactions}
\usage{
TraitMatching(
  B = 500,
  h_choice1 = NULL,
  h_choice2 = NULL,
  lambdas,
  betas,
  gammas,
  Us,
  Vs,
  Xs,
  Ws
)
}
\arguments{
\item{B}{Numeric. Number of times to perform the resampling. Default is 500.}

\item{h_choice1}{List of functions to be used for choosing h. These
functions should include functions that lead to randomness in the choice of
the component. If left NULL two choices will be considered: equal weight,
and proportional to the absolute value.}

\item{h_choice2}{List of functions to be used for choosing h. These
functions should include functions that lead to a single component. If left
NULL, the maximum absolute element will be chosen.}

\item{lambdas}{Matrix including the coefficients of the network model. The 
rows correspond to posterior samples and the columns to intercept and latent
factors components.}

\item{betas}{}

\item{Ws}{}
}
\description{
Using the posterior samples for covariate-informed latent factor network
model to infer trait imprtance in forming interactions.
}
