% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraitMatching2_function.R
\name{TraitMatching2}
\alias{TraitMatching2}
\title{Trait matching for species interactions}
\usage{
TraitMatching2(B = 500, mod_pL1s, Xs, Ws, obs_X, obs_W, obs_only = FALSE)
}
\arguments{
\item{B}{Numeric. Number of times to perform the resampling. Default is 500.}

\item{obs_only}{Logical. If set to TRUE only the observations with observed
covariate values will be used. Defaults to FALSE.}
}
\description{
Trait matching for species interactions
}
