% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraitMatching3_function.R
\name{TraitMatching3}
\alias{TraitMatching3}
\title{Trait matching for species interactions}
\usage{
TraitMatching3(B = 500, mod_pL1s, Xs, Ws, obs_X, obs_W)
}
\arguments{
\item{B}{Numeric. Number of times to perform the resampling. Default is 500.}
}
\description{
Trait matching for species interactions
}
