% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdOccur_function.R
\name{UpdOccur}
\alias{UpdOccur}
\title{Function that updates the indicator of occurence.}
\usage{
UpdOccur(detected, occur, probobs_curr, probobs_others, curr_inter, focus)
}
\arguments{
\item{detected}{Matrix. Rows are species, columns are studies. Indicator of
wether the species were detected to interact in the given study. Derived
from obs_A.}

\item{occur}{Matrix. Rows are species, columns are studies. It includes the
prior probabilities of occurence.}

\item{probobs_curr}{Vector of current values for the probability of
observation for the set of species we are updating.}

\item{probobs_others}{Vector of current values for the probability of
observation for the other set of species, the ones that are not being
updated.}

\item{curr_inter}{Matrix with 0,1 entries for the current posterior samples
of possible interactions. The rows correspond to the species we are updating
and the columns correspond to the other set of species.}

\item{focus}{Array of three dimensions corresponding to the two sets of
species, and the different studies. Values are 0 or 1. This array represents
whether the current study would be willing to record the interactions of a
given species. The value should be 1 except for studies that are animal or
plant-oriented, where the value for species that are not of focus will be 0.}
}
\description{
Each species might occur in the study area or not. Even though probabilities
of occurence can be provided, the detected interactions also inform us of
whether the species occur.
}
