% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdProbObs_function.R
\name{UpdProbObs}
\alias{UpdProbObs}
\title{Updating the species probability of being observed in a study}
\usage{
UpdProbObs(
  probobs_curr,
  probobs_others,
  curr_inter,
  obs_inter,
  mh_n,
  n_studies,
  latfac,
  coefs_probobs,
  var_probobs
)
}
\description{
A species probabiliy of being observed, pi, is updated based on the observed
interactions, latent traits, and current true interactions.
}
