% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdRho_function.R
\name{UpdRho}
\alias{UpdRho}
\title{Updating the correlation coefficient for latent factors.}
\usage{
UpdRho(curr_r, curr_S, corr_C, latfac, mh_n, prior_rho)
}
\arguments{
\item{curr_r}{Current value of the correlation coefficient.}

\item{curr_S}{Current value of the correlation matrix for the latent
factors.}

\item{corr_C}{Phylogenetic correlation matrix.}

\item{latfac}{Current values of the latent factors. Matrix with rows
corresponding to the species and columns to factors.}

\item{mh_n}{Parameter n in the Beta proposal for updating the 
correlation coefficient for latent factors.}

\item{prior_rho}{Hyperparameters of the beta prior for the weight rho in 
the correlation matrix for latent factors. Defaults to (5, 5).}
}
\description{
Updating the correlation coefficient for latent factors.
}
