% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdTraitCoef_function.R
\name{UpdTraitCoef}
\alias{UpdTraitCoef}
\title{Updating the coefficients in the traits model.}
\usage{
UpdTraitCoef(
  obs_cov,
  latfac,
  resid_var,
  shr_var,
  extra_var,
  curr_coefs,
  num_covs,
  prior_mu0,
  prior_sigmasq0
)
}
\arguments{
\item{obs_cov}{Matrix of observed values for the covariates. Rows
correspond to units and columns to covariates.}

\item{latfac}{Latent factor positions for all subjects. Matrix of rows
equal to the observations and columns equal to the number of factors.}

\item{resid_var}{Value of the residual variance for continuous traits.}

\item{shr_var}{Value of the variances across the H factor components.}

\item{extra_var}{Values of the additional variances of all coefficients.
Matrix of dimensions equal to the number of covariates and number of
factors.}

\item{curr_coefs}{Matrix of current coefficients. Dimensions correspond to
covariate and intercept/factor. They are used to draw Polya-Gamma latent
variables for the update of coefficients of binary models.}

\item{num_covs}{Vector including two values: number of continuous and binary
covariates.}

\item{prior_mu0}{Mean of the prior distribution on intercepts.}

\item{prior_sigmasq0}{Variance of the prior distribution on intercepts.}
}
\description{
Updating the coefficients in the traits model.
}
